[refactor] split swift-nio and url session based networking layer in there individual modules.
