+++ Sources/ElasticSwiftNetworkingUrlSession/SessionManager.swift
@@ -16,7 +16,7 @@ import ElasticSwiftCore
 /**
  Class maintaining URLSession for a Host
  */
-
+@available(iOS 10.0, macOS 10.10, tvOS 10, watchOS 3, *)
 class SessionManager: NSObject, URLSessionDelegate {
 
     private let logger = Logger(label: "org.pksprojects.ElasticSwift.Networking.SessionManager")
@@ -24,7 +24,6 @@ class SessionManager: NSObject, URLSessionDelegate {
     private var session: URLSession?
     private let url: URL
     
-    #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
     private var sslConfig: SSLConfiguration?
     
     init(forHost url: URL, sslConfig: SSLConfiguration? = nil) {
@@ -35,15 +34,7 @@ class SessionManager: NSObject, URLSessionDelegate {
         let queue = OperationQueue()
         self.session = URLSession(configuration: config, delegate: self, delegateQueue: queue)
     }
-    #else
-    init(forHost url: URL) {
-        self.url = url
-        super.init()
-        let config = URLSessionConfiguration.ephemeral
-        let queue = OperationQueue()
-        self.session = URLSession(configuration: config, delegate: self, delegateQueue: queue)
-    }
-    #endif
+    
     func createReqeust(_ httpRequest: HTTPRequest) -> URLRequest {
         var components =  URLComponents()
         components.queryItems = httpRequest.queryParams
@@ -83,9 +74,7 @@ class SessionManager: NSObject, URLSessionDelegate {
     }
 }
 
-
-#if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-// URLSession basic SSL support for apple platform
+@available(iOS 10.0, macOS 10.10, tvOS 10, watchOS 3, *)
 extension SessionManager {
     func urlSession(_ session: URLSession, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
         guard challenge.previousFailureCount == 0 else {
@@ -117,11 +106,8 @@ extension SessionManager {
     }
 }
 
-#endif
-
 // MARK:- Helper extension
-
-#if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
+@available(iOS 10.0, macOS 10.10, tvOS 10, watchOS 3, *)
 public extension SecCertificate {
     
     /**
@@ -165,4 +151,3 @@ public extension SecCertificate {
     }
     
 }
-#endif
